backend/pyproject.toml
[project]
name = "cms-backend"
version = "0.1.0"
description = "CMS Backend for Multi-Agent Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "redis[hiredis]>=5.0.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=43.0.0",
    "pytz>=2024.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
